requirements.txt
Django==2.1.4
djangorestframework==3.9.0
model-mommy==1.6.0
pytz==2018.7
six==1.12.0
